Create mathable module
