+++ lib/invoice_repository.rb
@@ -1,7 +1,9 @@
 require_relative 'sales_engine'
 require_relative 'invoice'
+require_relative 'mathable'
 
 class InvoiceRepository
+  include Mathable
   attr_reader :invoices
 
   def initialize(path, engine)
@@ -90,25 +92,6 @@ class InvoiceRepository
       hash[invoice.day_created] += 1
     end
     hash
-  end
-
-  def average(hash)
-    number_of_keys = hash.keys.count
-    total_invoices = all.count.to_f
-    (total_invoices / number_of_keys)
-  end
-
-  def hash_variance_from_mean(hash)
-    hash.each_with_object({}) do |original_hash, modified_hash|
-      modified_hash[original_hash.first] = (original_hash.last - average(hash)) ** 2
-    end
-  end
-
-  def standard_deviation(hash)
-    modified_hash = hash_variance_from_mean(hash)
-    divisor = hash.keys.count - 1
-    sum = modified_hash.values.sum
-    Math.sqrt(sum / divisor).round(2)
   end
 
   def top_sales_days
